Consolidate API transformers into API folder
