+++ app/Transformers/API/GameTransformer.php
@@ -17,7 +17,7 @@
  *    You should have received a copy of the GNU Affero General Public License
  *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
  */
-namespace App\Transformers\Multiplayer;
+namespace App\Transformers\API;
 
 use App\Models\Multiplayer\Game;
 use League\Fractal;
